Update the structure of the schemas in gs-agent (#10)
* update the structure of the schemas in gs-agent

* fix pre-commit
