--- src/agent/gs_agent/configs/__init__.py
@@ -1,17 +1,0 @@
-from .registry import (
-    PPO_DEFAULT,
-    PPO_GOAL_REACHING_MLP,
-    PPO_PENDULUM_MLP,
-    RUNNER_DEFAULT,
-    RUNNER_GOAL_REACHING_MLP,
-    RUNNER_PENDULUM_MLP,
-)
-
-__all__ = [
-    "PPO_DEFAULT",
-    "PPO_PENDULUM_MLP",
-    "PPO_GOAL_REACHING_MLP",
-    "RUNNER_DEFAULT",
-    "RUNNER_PENDULUM_MLP",
-    "RUNNER_GOAL_REACHING_MLP",
-]
